Fix cookie not being parsed on load.
If you have to stringify your object before saving, and you add that,
to the `_cookies` cache, then you'll return that on `load` after setting
it with `save`.

While if the value is already set, the script correctly parses the cookie
values and what you load is returned as an object.

This oneliner takes care of that for you.