Remove httpOnly from this library
Based on this comment:
https://github.com/reactivestack/cookies/issues/180#issuecomment-441851712

There's no way to create or modify an httpOnly cookie from the client
side. The fact that this option is available in `setCookie` is very
deceptive. The documentation makes it sound like you could set an
httpOnly cookie on the client side, to be sent with all future HTTP
requests.

Since there's no way to create an httpOnly cookie on the client side, we
can simply remove it from the documentation and code altogether.