Fix return type of decoded cookies
The type of cookies was always returning string even if a custom decode function was passed into the options variable. This change allows the return type to either be the type of the return from the decode function passed into the options or string in case the decode fails.

My initial intention was to fully type the cookie.parse function, but that is actually from another package not in this repo and that package is not typescript, so I think this is the best we get for now.